# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PromptHub API Documentation",
        "description": "API completa para gerenciamento de prompts de IA e agentes AI. Sistema que permite autenticacao via email/senha ou Google OAuth, cadastro e gerenciamento de agentes de IA personalizados, envio de prompts para OpenAI e Google Gemini.",
        "contact": {
            "name": "PromptHub Support",
            "email": "suporte@prompthub.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://dflix-prompthub-backend.a8zera.easypanel.host/api",
            "description": "Production Server (Easypanel)"
        },
        {
            "url": "https://api-prompthub.marcioleite.cloud/api",
            "description": "Production Server (Alternative)"
        }
    ],
    "paths": {
...
        "/ai/prompt": {
            "post": {
                "tags": [
                    "AI Prompts"
                ],
                "summary": "Enviar prompt para IA",
                "description": "Envia um prompt diretamente para o provedor de IA escolhido (OpenAI ou Google Gemini). Requer autenticacao e token da API do provedor.",
                "operationId": "sendPrompt",
                "requestBody": {
                    "description": "Dados do prompt para envio",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "provider",
                                    "api_token",
                                    "prompt"
                                ],
                                "properties": {
                                    "provider": {
                                        "description": "Provedor de IA a ser usado",
                                        "type": "string",
                                        "enum": [
                                            "openai",
                                            "gemini"
                                        ],
                                        "example": "openai"
                                    },
                                    "api_token": {
                                        "description": "Token de API do provedor escolhido",
                                        "type": "string",
                                        "example": "sk-proj-abc123..."
                                    },
                                    "prompt": {
                                        "description": "Texto do prompt (minimo 10 caracteres)",
                                        "type": "string",
                                        "minLength": 10,
                                        "example": "Explique o que e inteligencia artificial em 100 palavras"
                                    },
                                    "temperature": {
                                        "description": "Controle de criatividade (0=deterministico, 1=criativo)",
                                        "type": "number",
                                        "format": "float",
                                        "maximum": 1,
                                        "minimum": 0,
                                        "example": 0.7
                                    },
                                    "max_tokens": {
                                        "description": "Limite maximo de tokens na resposta",
                                        "type": "integer",
                                        "maximum": 4000,
                                        "minimum": 100,
                                        "example": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta da IA recebida com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "response": {
                                            "type": "string",
                                            "example": "Inteligencia artificial e a simulacao de processos de inteligencia humana por sistemas computacionais..."
                                        },
                                        "provider": {
                                            "type": "string",
                                            "example": "openai"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validacao",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The provider field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao processar prompt",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao processar prompt: OpenAI API Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nao autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}